structure of the backend library and tests that are currently failing in one specific situation (with the LayerTreeNode problem QgsLayerTreeNode #514
