+++ QgisModelBaker/internal_libs/toppingmaker/utils.py
@@ -17,22 +17,8 @@
  *                                                                         *
  ***************************************************************************/
 """
-import os
 import re
-import shutil
 import unicodedata
-
-from .target import Target
-
-
-def toppingfile_link(target: Target, type: str, path: str):
-    filename_slug = f"{slugify(target.projectname)}_{os.path.basename(path)}"
-    absolute_filedir_path, relative_filedir_path = target.filedir_path(type)
-    shutil.copy(
-        path,
-        os.path.join(absolute_filedir_path, filename_slug),
-    )
-    return target.path_resolver(target, filename_slug, type)
 
 
 def slugify(text: str) -> str:
